Sprint 1: land harness and headless lifecycle parity
Acceptance checks verified:
- port the harness FileSystem, execution environment, SessionStorage, SessionRepo, memory/JSONL repositories, UUID generation, and rehydrate-from-bytes path
- preserve legacy, future, malformed-line, delayed-flush, unknown-member, and session replacement quirks from pinned TypeScript
- route print, JSON, and RPC through one replaceable runtime with extension lifecycle and event subscription rebinding
- generate and verify the upstream-derived F6Harness corpus; keep all existing fixtures byte-clean
- wire the credential-gated OpenAI/Anthropic nightly live corpus

Verification:
- make build
- make test
- make lint
- make fixtures-check
- make upstream-rpc-tests (6 files, 27 tests)
- CGO_ENABLED=0 cross-builds for linux/darwin amd64/arm64
- go mod verify; no go.mod or go.sum delta
- harness non-test LOC 5,074 vs upstream 6,032 (0.84x)

+++ codingagent/session_runtime_test.go
@@ -830,8 +830,8 @@ func TestLongFauxSessionCompactsAtHarnessBoundary(t *testing.T) {
 		t.Fatal(err)
 	}
 	compaction := branch[len(branch)-1]
-	if want == nil || compaction.FirstKeptEntryID != want.FirstKeptEntryID || compaction.TokensBefore != want.TokensBefore {
-		t.Fatalf("compaction boundary=(%s,%d), want=(%#v)", compaction.FirstKeptEntryID, compaction.TokensBefore, want)
+	if want == nil || compaction.FirstKeptEntryID != want.FirstKeptEntryID || compaction.TokensBefore != float64(want.TokensBefore) {
+		t.Fatalf("compaction boundary=(%s,%g), want=(%#v)", compaction.FirstKeptEntryID, compaction.TokensBefore, want)
 	}
 	usage := runtime.GetContextUsage()
 	if usage == nil || usage.Tokens != nil || usage.Percent != nil {
